+++ ChangeLog.md
@@ -98,6 +98,7 @@ Nazara Engine:
 - ⚠️ Refactored Mesh/SubMesh, allowing a submesh to be attached to multiple meshes, deprecating Create/Destroy methods
 - SubMesh class now has a OnSubMeshInvalidateAABB signal, triggered when a new AABB is set to the submesh
 - Mesh class now has a OnMeshInvalidateAABB signal, triggered when a mesh invalidates its AABB, which is also submesh updates its AABB
+- Model now invalidate properly their bounding volume when their mesh AABB is updated
 
 Nazara Development Kit:
 - Added ImageWidget (#139)